tink/backend/skill_quest/genai_client.py
"""Lazy-initialized GenAI client to avoid import-time API key errors."""

from google import genai

_client: genai.Client | None = None


def get_client() -> genai.Client:
    global _client
    if _client is None:
        _client = genai.Client()
    return _client
